Move helpers from about page tests to base
